REF: Typescript refactor.
