+++ src/components/modules/Header/StyledHeader.tsx
@@ -1,6 +1,6 @@
 // Library Imports
 import styled from 'styled-components'
-import media from 'globalConfig/mobile'
+import { mediaQueries } from 'globalConfig/mobile'
 
 const StyledHeader = styled.section`
 /* MOBILE FIRST DEFAULT STYLES */
@@ -8,38 +8,37 @@ const StyledHeader = styled.section`
   width: 100%;
 
 
-/* Min width of 375 */
-  ${media.iphoneSe `
-    
+  /* Min width of 375 */
+  ${mediaQueries('iphoneSe')`
   `}
 
 /* Min width of 414 */
-  ${media.iphoneXr `
+  ${mediaQueries('iphoneXr') `
     
   `}
 
 /* Min width of 576 */
-  ${media.landscapePhones `
+  ${mediaQueries('landscapePhones') `
     
   `}
 
 /* Min width of 768 */
-  ${media.tablet `
+  ${mediaQueries('tablet') `
     
   `}
 
 /* Min width of 1024 */
-  ${media.desktop `
+  ${mediaQueries('desktop') `
     
   `}
 
 /* Min width of 1200 */
-  ${media.widescreen `
+  ${mediaQueries('widescreen') `
     
   `}
 
 /* Min width of 1980 */
-  ${media.udh `
+  ${mediaQueries('udh') `
     
   `}
 `
